Refactor: 共通化
